services/ai-orchestrator/pyproject.toml
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "healthcare-ai-orchestrator"
version = "1.0.0"
description = "AI orchestration service for test generation and RAG"
authors = ["Healthcare Compliance Team"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.104.1"
uvicorn = {extras = ["standard"], version = "^0.24.0"}
pydantic = "^2.5.0"
pydantic-settings = "^2.1.0"
google-cloud-aiplatform = "^1.38.0"
google-cloud-storage = "^2.10.0"
google-cloud-bigquery = "^3.13.0"
google-cloud-pubsub = "^2.18.4"
google-cloud-secret-manager = "^2.17.0"
google-cloud-logging = "^3.8.0"
google-cloud-trace = "^1.12.0"
google-auth = "^2.23.4"
firebase-admin = "^6.2.0"
httpx = "^0.25.2"
structlog = "^23.2.0"
tenacity = "^8.2.3"
numpy = "^1.24.4"
sentence-transformers = "^2.2.2"
langchain = "^0.0.350"
langchain-google-vertexai = "^0.0.6"
tiktoken = "^0.5.2"
chromadb = "^0.4.18"
jinja2 = "^3.1.2"
aiofiles = "^23.2.1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-asyncio = "^0.21.1"
pytest-cov = "^4.1.0"
black = "^23.11.0"
isort = "^5.12.0"
flake8 = "^6.1.0"
mypy = "^1.7.1"
httpx = "^0.25.2"

[tool.black]
line-length = 88
target-version = ['py311']

[tool.isort]
profile = "black"
multi_line_output = 3

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
